feat: added performance profile and some cleanup
