+++ pkg/evaluate.go
@@ -1,67 +1,5 @@
 package libra
 
-// Piece-Square Tables (simplified, values in centipawns)
-var pawnPST = [64]int{
-	0, 0, 0, 0, 0, 0, 0, 0,
-	50, 50, 50, 50, 50, 50, 50, 50,
-	10, 10, 20, 30, 30, 20, 10, 10,
-	5, 5, 10, 25, 25, 10, 5, 5,
-	0, 0, 0, 20, 20, 0, 0, 0,
-	5, -5, -10, 0, 0, -10, -5, 5,
-	5, 10, 10, -20, -20, 10, 10, 5,
-	0, 0, 0, 0, 0, 0, 0, 0,
-}
-var knightPST = [64]int{
-	-50, -40, -30, -30, -30, -30, -40, -50,
-	-40, -20, 0, 0, 0, 0, -20, -40,
-	-30, 0, 10, 15, 15, 10, 0, -30,
-	-30, 5, 15, 20, 20, 15, 5, -30,
-	-30, 0, 15, 20, 20, 15, 0, -30,
-	-30, 5, 10, 15, 15, 10, 5, -30,
-	-40, -20, 0, 5, 5, 0, -20, -40,
-	-50, -40, -30, -30, -30, -30, -40, -50,
-}
-var bishopPST = [64]int{
-	-20, -10, -10, -10, -10, -10, -10, -20,
-	-10, 5, 0, 0, 0, 0, 5, -10,
-	-10, 10, 10, 10, 10, 10, 10, -10,
-	-10, 0, 10, 10, 10, 10, 0, -10,
-	-10, 5, 5, 10, 10, 5, 5, -10,
-	-10, 0, 5, 10, 10, 5, 0, -10,
-	-10, 0, 0, 0, 0, 0, 0, -10,
-	-20, -10, -10, -10, -10, -10, -10, -20,
-}
-var rookPST = [64]int{
-	0, 0, 0, 0, 0, 0, 0, 0,
-	5, 10, 10, 10, 10, 10, 10, 5,
-	-5, 0, 0, 0, 0, 0, 0, -5,
-	-5, 0, 0, 0, 0, 0, 0, -5,
-	-5, 0, 0, 0, 0, 0, 0, -5,
-	-5, 0, 0, 0, 0, 0, 0, -5,
-	-5, 0, 0, 0, 0, 0, 0, -5,
-	0, 0, 0, 5, 5, 0, 0, 0,
-}
-var queenPST = [64]int{
-	-20, -10, -10, -5, -5, -10, -10, -20,
-	-10, 0, 0, 0, 0, 0, 0, -10,
-	-10, 0, 5, 5, 5, 5, 0, -10,
-	-5, 0, 5, 5, 5, 5, 0, -5,
-	0, 0, 5, 5, 5, 5, 0, -5,
-	-10, 5, 5, 5, 5, 5, 0, -10,
-	-10, 0, 5, 0, 0, 0, 0, -10,
-	-20, -10, -10, -5, -5, -10, -10, -20,
-}
-var kingPST = [64]int{
-	-30, -40, -40, -50, -50, -40, -40, -30,
-	-30, -40, -40, -50, -50, -40, -40, -30,
-	-30, -40, -40, -50, -50, -40, -40, -30,
-	-30, -40, -40, -50, -50, -40, -40, -30,
-	-20, -30, -30, -40, -40, -30, -30, -20,
-	-10, -20, -20, -20, -20, -20, -20, -10,
-	20, 20, 0, 0, 0, 0, 20, 20,
-	20, 30, 10, 0, 0, 10, 30, 20,
-}
-
 // mirrorIndex mirrors a square index for black's perspective
 func mirrorIndex(idx byte) byte {
 	return 56 ^ (idx & 56) | (idx & 7)
