+++ pkg/move.go
@@ -40,6 +40,17 @@ func NewMove(from byte, to byte, moveType byte, data [2]byte) Move {
 	}
 }
 
+type MoveState struct {
+	Position             [64]byte
+	AttackedSquares      [64]bool
+	Pieces               PieceColorLocation
+	CastlingAvailability CastlingAvailability
+	OnPassant            byte
+	HalfMoveClock        int
+	FullMoveCounter      int
+	WhiteToMove          bool
+}
+
 func generateSquaresToEdge() [64][8]byte {
 	squares := [64][8]byte{}
 	for i := range squares {
@@ -123,17 +134,8 @@ var SquaresToEdge [64][8]byte = generateSquaresToEdge()
 var SquareKnightJumps [64][8]byte = generateKnightJumps()
 var BoardDirOffsets [8]int8 = [8]int8{-8, 1, 8, -1, -7, 9, 7, -9}
 
-type MoveState struct {
-	Position             [64]byte
-	CastlingAvailability CastlingAvailability
-	OnPassant            byte
-	HalfMoveClock        int
-	FullMoveCounter      int
-	WhiteToMove          bool
-}
-
-// MakeMove now returns a MoveState for undoing the move
-func (board *Board) MakeMove(move Move) MoveState {
+// Move returns a MoveState for undoing the move
+func (board *Board) Move(move Move) MoveState {
 	// Save current state
 	prev := MoveState{
 		Position:             board.Position,
@@ -142,6 +144,8 @@ func (board *Board) MakeMove(move Move) MoveState {
 		HalfMoveClock:        board.HalfMoveClock,
 		FullMoveCounter:      board.FullMoveCounter,
 		WhiteToMove:          board.WhiteToMove,
+		AttackedSquares:      board.AttackedSquares,
+		Pieces:               board.Pieces,
 	}
 
 	if !board.WhiteToMove {
@@ -274,5 +278,6 @@ func (board *Board) UndoMove(state MoveState) {
 	board.HalfMoveClock = state.HalfMoveClock
 	board.FullMoveCounter = state.FullMoveCounter
 	board.WhiteToMove = state.WhiteToMove
-	board.UpdatePiecesLocation()
+	board.AttackedSquares = state.AttackedSquares
+	board.Pieces = state.Pieces
 }
